Hardcoded screen art depending of  pod_id

+++ src/A50_Cellmask.h
@@ -1,17 +1,17 @@
 class Cellmask {
 
   public:
-    Cellmask(byte h, byte s, byte b){
+    Cellmask(int h, byte s, byte b){
       _h = h;
       _s = s;
       _b = b;
     }
-    byte getHue() {return _h;}
+    int getHue() {return _h;}
     byte getSaturation() {return _s;}
     byte getBright() {return _b;}
 
   private:
-    byte _h;
+    int _h;
     byte _s;
     byte _b;
